Main addition of this commit was to add functionality to generate the `ActionChoice`s of both teams in order to give it to the UI renderer. In doing so a number of other changes were made:
Renamed the `BattlerNumber` and `MoveNumber`
variants.

Removed the `ActionChoice::None` variant, and replaced usage with
`Option<ActionChoice>`.

Commented out some out-of-date functions that generated hardcoded user
input for the engine. As such the ability to advance the simulator with
the N-key has been removed.

Added a method to get the uids for thes moves on a `Battler` and a
method on `BattlerTeam` to get the active_battler on that team.