Implemented stat modifieres, raising and lowering stat actions, used this to implement `Growl`.
renamed all `context` bindings to `ctx` to reduce visual clutter. The
struct is so central to the engine that I think it will be
recognisable if  it is used consistently. Also renamed `Lcrng` to
`Prng` for clarity.

Also added `push_message()` and `push_messages()` functions to the API
to make it more concise.