Refactored app further. Renamed a bunch of items and reordered them.
Changed the way message log scrolling works to allow for smarter
scrolling behaviour (snaps to next turn and prevents scrolling below
the beginning of the last turn).

Changed `team_name` on `TeamUiState` to be a `TeamID` for strong typing.

Added an invalid choice message in case the simulator ever fails to read
user input.